+++ RELEASE_NOTES.md
@@ -2,4 +2,4 @@
 
 Les Co'Pintes évolue avec une nouvelle version.
 
-- Modification de la reidrection vers l'application ASDRO
+- Ajout de l'évènement au calendrier personnel